replace the door kicker with mqtt
The previous implementation lacks the support of generating events when the status changes..
The root cause is the bot doesn't know if it's open or not, it has to check timestamp
when asked.
In preparation for an external http:// space api and picture service, the
bot needs to create an event when the door status changes.